Added support for "all" events watching.
Had to add a makeshift file with all events listed out, but it works.
Also, removing an event result from the results list removes it from
watching.
Also, the results pane tries to stay on the bottom, if it was scrolled
there, for ease of watching events in real time.